Revert "Merge pull request #472 from urbanairship/MB-401"
This reverts commit cef9dfbf39a255652ef9f8e101207dfe08d14d4b, reversing
changes made to 8612ebf7e554c6978dc25fc21a315a40555b5647.

Conflicts:
	Airship/Common/UAActionJSDelegate.m
	Airship/Common/UACustomEvent.h
	Airship/Common/UACustomEvent.m
	Airship/Common/UAEvent.h
	Airship/Common/UALocationService.m

+++ Airship/Common/UAEvent.h
@@ -24,12 +24,15 @@
  */
 
 
-#define kEventAppInitSize               450//397 w/ push id, no inbox id
-#define kEventAppExitSize               200//136 w/ only network type
-#define kEventDeviceRegistrationSize    200//153 w/ only user info
-#define kEventPushReceivedSize          200//160 w/ uuid push info
-#define kEventAppActiveSize             120
-#define kEventAppInactiveSize           120
+
+#define kEventAppInitSize               (NSUInteger) 450//397 w/ push id, no inbox id
+#define kEventAppExitSize               (NSUInteger) 200//136 w/ only network type
+
+#define kEventDeviceRegistrationSize    (NSUInteger) 200//153 w/ only user info
+#define kEventPushReceivedSize          (NSUInteger) 200//160 w/ uuid push info
+#define kEventAppActiveSize             (NSUInteger) 120
+#define kEventAppInactiveSize           (NSUInteger) 120
+
 
 @interface UAEvent : NSObject
